Work around RGBXYZ stream hangs (issue #25)
Apparently, this is triggered when the depth_image_proc/register nodelet
fails with a TF2 'timestamp is in the future' error and doesn't publish
the registered depth image properly. This seems to cause the ROS
Synchronizer filter to choke.

Work around by moving the published timestamp further into the future.