Clean up system process calls
If the ROS node aborted before the system calls completed, orphan
processes could be left on the system. This was found in such cases
as the node aborting due to missing calibration data.

Switch to Fork and exec to ensure proper clean up of processes.
Keep a queue of forked processes and signal on exit.