In coordinate system transforms, fix which extrincits we use and use matrix properly.
Two bugs which cancel out each other for rotation, but not translation:
- it seems that ROS and Realsense use different conventions of coordinate
system transformations. In ROS, it is defined as a transformation of child
fame coordinates to parent frame coordinates (see
http://wiki.ros.org/tf/Overview/Transformations), while in RealSense
it seems to be transformation of "from" frame coordinates to "to" frame
coordinates. Thus, the order needs to be reversed.
- the matrix in RealSense extrinsics is stored in column-major format, while
Eigen::Matrix3f expects row-major, causing the matrix to be transposed.

To see that this is a problem, one can open rviz and add the pointcloud and the
color/image_raw camera. From the camera viewpoint, the images should align, but
don't. This patch doesn't yet solve the whole problem, but makes it smaller.