Estimate accurate timestamps by calculating clock offset
The device clock can advance at the different speed than the system
clock, which causes the systematic drifts of timestamps. This change
allows to estimate the clock offset using a low pass filter, and
compensate for the drift. Experimental results show that the typical
drift is about 0.000075 sec/s, equivalent to 1 frame different in 6
minutes at 30Hz. With this change the drift is successfully removed.