Add option to use external nodelet manager
Right now a nodelet manager is always created when running the realsense
node. This is not always a wanted behaviour as a nodelet manager might
already be running somewhere else in the ROS system. This patch adds a
parameter to disable launching a new nodelet manager. In this case the
nodelet manager must be provided using the manager arg. The new
parameter is set to false by default, so existing behaviour is not
changed.