Add support for rgb_camera.power_line_frequency
When running with LibRealSense v2.54.0 (built from source) a warning is issued, because the value of rgb_camera.power_line_frequency defaults to "{rgb_camera.power_line_frequency}". which is not in the range [0,2].

Adding this line allows for the rgb_camera.power_line_frequency to be set as a parameter.